+++ src/App.tsx
@@ -1,0 +1,5 @@
+function App() {
+  return <div className='App'>LINKFIT 시작</div>;
+}
+
+export default App;
